Abstract HDKeyring and move ETH specific codes into EthereumKeyring
