+++ components/brave_wallet/browser/test/BUILD.gn
@@ -25,10 +25,10 @@ source_set("brave_wallet_unit_tests") {
     "//brave/components/brave_wallet/browser/eth_response_parser_unittest.cc",
     "//brave/components/brave_wallet/browser/eth_transaction_unittest.cc",
     "//brave/components/brave_wallet/browser/eth_tx_service_unittest.cc",
+    "//brave/components/brave_wallet/browser/ethereum_keyring_unittest.cc",
     "//brave/components/brave_wallet/browser/ethereum_permission_utils_unittest.cc",
     "//brave/components/brave_wallet/browser/fil_requests_unittest.cc",
     "//brave/components/brave_wallet/browser/fil_response_parser_unittest.cc",
-    "//brave/components/brave_wallet/browser/hd_keyring_unittest.cc",
     "//brave/components/brave_wallet/browser/internal/hd_key_ed25519_unittest.cc",
     "//brave/components/brave_wallet/browser/internal/hd_key_unittest.cc",
     "//brave/components/brave_wallet/browser/json_rpc_response_parser_unittest.cc",
